+++ lib/basic/string/strtol.c
@@ -22,8 +22,8 @@
 #include <string.h>
 #include <assert.h>
 
-s32 strtol(const char *nptr,char **endptr,s32 base) {
-	s32 i = 0;
+long strtol(const char *nptr,char **endptr,int base) {
+	long i = 0;
 	bool neg = false;
 	char c;
 	assert(nptr);
